fix: type warning
