--- public/assets/index.898c63c1.js
@@ -1,49 +1,0 @@
-var re=Object.defineProperty;var A=Object.getOwnPropertySymbols;var le=Object.prototype.hasOwnProperty,ae=Object.prototype.propertyIsEnumerable;var B=(e,t,o)=>t in e?re(e,t,{enumerable:!0,configurable:!0,writable:!0,value:o}):e[t]=o,f=(e,t)=>{for(var o in t||(t={}))le.call(t,o)&&B(e,o,t[o]);if(A)for(var o of A(t))ae.call(t,o)&&B(e,o,t[o]);return e};import{d as m,r as D,c as u,a as s,b as d,w as g,n,o as i,e as w,g as I,f as ue,h,i as S,F as b,j as k,t as _,k as M,u as ie,l as F,m as q,p as ce,q as pe,s as de,v as _e}from"./vendor.b4423c4d.js";const me=function(){const t=document.createElement("link").relList;if(t&&t.supports&&t.supports("modulepreload"))return;for(const r of document.querySelectorAll('link[rel="modulepreload"]'))c(r);new MutationObserver(r=>{for(const a of r)if(a.type==="childList")for(const l of a.addedNodes)l.tagName==="LINK"&&l.rel==="modulepreload"&&c(l)}).observe(document,{childList:!0,subtree:!0});function o(r){const a={};return r.integrity&&(a.integrity=r.integrity),r.referrerpolicy&&(a.referrerPolicy=r.referrerpolicy),r.crossorigin==="use-credentials"?a.credentials="include":r.crossorigin==="anonymous"?a.credentials="omit":a.credentials="same-origin",a}function c(r){if(r.ep)return;r.ep=!0;const a=o(r);fetch(r.href,a)}};me();const $e="_wrapper_1v68l_20",ye="_headerContents_1v68l_26",he="_logo_1v68l_37",fe="_rightContents_1v68l_43",ge="_list_1v68l_48",be="_link_1v68l_59";var ve={wrapper:$e,headerContents:ye,logo:he,rightContents:fe,list:ge,link:be},$=(e,t)=>{for(const[o,c]of t)e[o]=c;return e};const we=m({setup(){return{}}}),ke=w("R - World"),Me=w("home"),Ce=w("member"),De=w("signup");function Ee(e,t,o,c,r,a){const l=D("router-link");return i(),u("header",{class:n(e.$style.wrapper)},[s("div",{class:n(e.$style.headerContents)},[d(l,{to:"/",class:n(e.$style.logo)},{default:g(()=>[ke]),_:1},8,["class"]),s("div",{class:n(e.$style.rightContents)},[s("ul",{class:n(e.$style.list)},[s("li",null,[d(l,{to:"/",class:n(e.$style.link)},{default:g(()=>[Me]),_:1},8,["class"])]),s("li",null,[d(l,{to:"/member",class:n(e.$style.link)},{default:g(()=>[Ce]),_:1},8,["class"])]),s("li",null,[d(l,{to:"/signup",class:n(e.$style.link)},{default:g(()=>[De]),_:1},8,["class"])])],2)],2)],2)],2)}const L={};L.$style=ve;var E=$(we,[["render",Ee],["__cssModules",L]]);const Ae="_wrapper_1i1bc_20",Be="_hello_1i1bc_25",Ie="_container_1i1bc_33",Se="_section_1i1bc_46";var Fe={wrapper:Ae,hello:Be,container:Ie,section:Se};const qe=m({components:{OGlobalHeader:E},setup(){return{}}}),Le=s("h1",null,"R - World",-1),Ne=s("p",null,"Let them draw big dreams. Make big dreams come true.",-1),Oe=[Le,Ne],Pe=s("h2",null,"Mission",-1),Re=s("p",null,"\u6311\u6226\u8005\u306B\u81EA\u7531\u3067\u5927\u304D\u306A\u5922\u3092\u63CF\u304B\u305B\u3001\u305D\u308C\u3092\u5B9F\u73FE\u3055\u305B\u308B",-1),Ve=s("h3",null,"Organization",-1),ze=s("p",null,"R\u3061\u3083\u3093\u304C\u51FA\u4F1A\u3063\u3066\u304D\u305F\u6109\u5FEB\u306A\u66F2\u8005\u305F\u3061\u306E\u96C6\u307E\u308A",-1),Ge=[Pe,Re,Ve,ze];function He(e,t,o,c,r,a){return i(),u("main",{class:n(e.$style.wrapper)},[s("div",{class:n(e.$style.hello)},[s("div",{class:n(e.$style.container)},Oe,2)],2),s("section",{class:n(e.$style.section)},Ge,2)],2)}const N={};N.$style=Fe;var Te=$(qe,[["render",He],["__cssModules",N]]);const Ue=I`
-    query getMember($id: String!) {
-  getMember(id: $id) {
-    id
-    updatedAt
-    name
-    occupation
-    self_introduction
-    company
-    experience_year
-    profile_image {
-      url
-      height
-      width
-    }
-    sns {
-      fieldId
-      facebook
-      twitter
-      github
-    }
-    experience {
-      fieldId
-      company
-      from
-      describe
-      skills
-      occupation
-    }
-    skills
-  }
-}
-    `,je=I`
-    query getMembers {
-  getMembers {
-    id
-    updatedAt
-    name
-    occupation
-    company
-    experience_year
-    profile_image {
-      url
-      height
-      width
-    }
-  }
-}
-    `,We=(e,t)=>e();function Ke(e,t=We){return{getMember(o,c){return t(r=>e.request(Ue,o,f(f({},c),r)),"getMember")},getMembers(o,c){return t(r=>e.request(je,o,f(f({},c),r)),"getMembers")}}}const Qe=new ue.GraphQLClient("https://toipptakosan11-71185.web.app/api/graphql"),O=Ke(Qe);async function Xe(e){const t=await O.getMember({id:e});return console.log(t.getMember),t.getMember}async function Ye(){const e=await O.getMembers();return console.log(e.getMembers),e.getMembers}function P(){const e=h(),t=h();return{member:e,getMember:Xe,members:t,getMembers:Ye}}const Je="_wrapper_1mfyn_20",Ze="_card_1mfyn_31",xe="_cardContent_1mfyn_35",es="_profileImage_1mfyn_45",ss="_name_1mfyn_51";var ts={wrapper:Je,card:Ze,cardContent:xe,profileImage:es,name:ss};const ns=m({setup(){const{members:e,getMembers:t}=P();return S(async()=>{e.value=await t()}),{members:e}}}),os=s("h1",null,"Member",-1),rs=["src"];function ls(e,t,o,c,r,a){const l=D("router-link");return i(),u("div",{class:n(e.$style.wrapper)},[os,s("ul",null,[(i(!0),u(b,null,k(e.members,({id:p,updatedAt:y,name:C,occupation:v,company:te,experience_year:ne,profile_image:oe})=>(i(),u("li",{key:p,class:n(e.$style.card)},[d(l,{to:`/member/${p}`,class:n(e.$style.cardContent)},{default:g(()=>[s("img",{src:oe.url,class:n(e.$style.profileImage)},null,10,rs),s("div",null,[s("div",{class:n(e.$style.name)},_(C),3),s("ul",null,[s("li",null,"\u8077\u7A2E: "+_(v[0]),1),s("li",null,"\u7D4C\u9A13\u5E74\u6570: "+_(ne)+"\u5E74",1),s("li",null,"\u52E4\u52D9\u5148: "+_(te),1),s("li",null,"\u6700\u7D42\u66F4\u65B0\u65E5: "+_(y),1)])])]),_:2},1032,["to","class"])],2))),128))])],2)}const R={};R.$style=ts;var as=$(ns,[["render",ls],["__cssModules",R]]);const us="_wrapper_1r3c5_20",is="_tags_1r3c5_28",cs="_tag_1r3c5_28";var ps={wrapper:us,tags:is,tag:cs};const ds=m({props:{skills:{type:Array,required:!0}},setup(){return{}}}),_s=s("div",null,"Skills",-1);function ms(e,t,o,c,r,a){return i(),u("div",{class:n(e.$style.wrapper)},[_s,s("ul",{class:n(e.$style.tags)},[(i(!0),u(b,null,k(e.skills,l=>(i(),u("li",{key:l},[s("span",{class:n(e.$style.tag)},_(l),3)]))),128))],2)],2)}const V={};V.$style=ps;var z=$(ds,[["render",ms],["__cssModules",V]]);const $s="_wrapper_hf545_20",ys="_contents_hf545_28",hs="_cardBox_hf545_35",fs="_point_hf545_39",gs="_cardContents_hf545_53",bs="_companyName_hf545_58",vs="_companyPosition_hf545_62",ws="_companyDescription_hf545_67",ks="_companySkills_hf545_74";var Ms={wrapper:$s,contents:ys,cardBox:hs,point:fs,cardContents:gs,companyName:bs,companyPosition:vs,companyDescription:ws,companySkills:ks};const Cs=m({props:{member:{type:Object,required:!0}},setup(){return{}}}),Ds=s("div",null,"Experience",-1);function Es(e,t,o,c,r,a){return i(),u("section",{class:n(e.$style.wrapper)},[Ds,s("div",{class:n(e.$style.contents)},[s("ul",null,[(i(!0),u(b,null,k(e.member.experience,({company:l,describe:p,occupation:y,skills:C})=>(i(),u("li",{key:l},[s("div",{class:n(e.$style.point)},null,2),s("div",{class:n(e.$style.cardBox)},[s("div",{class:n(e.$style.cardContents)},[s("div",{class:n(e.$style.companyName)},_(l),3),s("div",{class:n(e.$style.companyPosition)},_(y[0]),3),s("div",{class:n(e.$style.companyDescription)},_(p),3),s("ul",{class:n(e.$style.companySkills)},[(i(!0),u(b,null,k(C,v=>(i(),u("li",{key:`${l}-${v}`},[s("span",null,_(v),1)]))),128))],2)],2)],2)]))),128))])],2)],2)}const G={};G.$style=Ms;var H=$(Cs,[["render",Es],["__cssModules",G]]),As="./assets/facebook-brands.e4e4ffee.svg",Bs="./assets/twitter-brands.dd81ac6e.svg",Is="./assets/github-brands.4ea99f48.svg";const Ss="_wrapper_1yvxw_20",Fs="_sns_1yvxw_28",qs="_icon_1yvxw_32";var Ls={wrapper:Ss,sns:Fs,icon:qs};const Ns=m({props:{member:{type:Object,required:!0}},setup(){return{}}}),Os=s("div",null,"Self Introduction",-1),Ps={key:0},Rs=["href"],Vs={key:1},zs=["href"],Gs={key:2},Hs=["href"];function Ts(e,t,o,c,r,a){return i(),u("section",{class:n(e.$style.wrapper)},[Os,s("p",null,_(e.member.self_introduction),1),s("ul",{class:n(e.$style.sns)},[e.member.sns.facebook?(i(),u("li",Ps,[s("a",{href:e.member.sns.facebook},[s("img",{src:As,alt:"facebook icon",class:n(e.$style.icon)},null,2)],8,Rs)])):M("",!0),e.member.sns.twitter?(i(),u("li",Vs,[s("a",{href:e.member.sns.twitter},[s("img",{src:Bs,alt:"twitter icon",class:n(e.$style.icon)},null,2)],8,zs)])):M("",!0),e.member.sns.github?(i(),u("li",Gs,[s("a",{href:e.member.sns.github},[s("img",{src:Is,alt:"github icon",class:n(e.$style.icon)},null,2)],8,Hs)])):M("",!0)],2)],2)}const T={};T.$style=Ls;var U=$(Ns,[["render",Ts],["__cssModules",T]]);const Us="_wrapper_1fg8u_20",js="_header_1fg8u_26",Ws="_headerContent_1fg8u_35",Ks="_rightContents_1fg8u_51",Qs="_main_1fg8u_63";var Xs={wrapper:Us,header:js,headerContent:Ws,rightContents:Ks,main:Qs};const Ys=m({components:{OMemberDetailSelfIntroduction:U,OMemberDetailExperience:H,OMemberDetailSkills:z},setup(){const e=ie(),{member:t,getMember:o}=P();return S(async()=>{t.value=await o(e.params.id)}),{member:t}}}),Js=["src"];function Zs(e,t,o,c,r,a){const l=U,p=H,y=z;return e.member?(i(),u("div",{key:0,class:n(e.$style.wrapper)},[s("div",{class:n(e.$style.header)},[s("div",{class:n(e.$style.headerContent)},[s("img",{src:e.member.profile_image.url,alt:"profile image"},null,8,Js),s("div",{class:n(e.$style.rightContents)},[s("h3",null,_(e.member.name),1),s("div",null,_(e.member.company)+" / "+_(e.member.occupation[0]),1),s("div",null,"ID: "+_(e.member.id),1)],2)],2)],2),s("main",{class:n(e.$style.main)},[d(l,{member:e.member},null,8,["member"]),d(p,{member:e.member},null,8,["member"]),d(y,{skills:e.member.skills},null,8,["skills"])],2)],2)):M("",!0)}const j={};j.$style=Xs;var xs=$(Ys,[["render",Zs],["__cssModules",j]]);const et="_button_177zs_20";var st={button:et};const tt=m({props:{text:{type:String,required:!0},onClick:{type:Function,required:!0}}});function nt(e,t,o,c,r,a){return i(),u("button",{class:n(e.$style.button),onClick:t[0]||(t[0]=()=>e.onClick)},_(e.text),3)}const W={};W.$style=st;var ot=$(tt,[["render",nt],["__cssModules",W]]);const rt="_input_2ad87_20";var lt={input:rt};const at=m({props:{type:{type:String,default:"text"},placeholder:{type:String,default:""},modelValue:{type:String,required:!0}},emits:["update:modelValue"]}),ut=["type","placeholder","value"];function it(e,t,o,c,r,a){return i(),u("input",{type:e.type,placeholder:e.placeholder,class:n(e.$style.input),value:e.modelValue,onInput:t[0]||(t[0]=l=>e.$emit("update:modelValue",l.currentTarget))},null,42,ut)}const K={};K.$style=lt;var ct=$(at,[["render",it],["__cssModules",K]]);const pt="_wrapper_1ycfq_20",dt="_content_1ycfq_27",_t="_button_1ycfq_32";var mt={wrapper:pt,content:dt,button:_t};const $t=m({props:{email:{type:String,required:!0},password:{type:String,required:!0},buttonText:{type:String,required:!0},onClick:Function},emits:["update:email","update:password"]});function yt(e,t,o,c,r,a){const l=ct,p=ot;return i(),u("div",{class:n(e.$style.wrapper)},[d(l,{type:"text",placeholder:"Email",class:n(e.$style.content),value:e.email,onInput:t[0]||(t[0]=y=>e.$emit("update:email",y.target.value))},null,8,["class","value"]),d(l,{type:"password",placeholder:"Password",class:n(e.$style.content),value:e.password,onInput:t[1]||(t[1]=y=>e.$emit("update:password",y.target.value))},null,8,["class","value"]),d(p,{text:e.buttonText,onClick:e.onClick,class:n(e.$style.button)},null,8,["text","onClick","class"])],2)}const Q={};Q.$style=mt;var X=$($t,[["render",yt],["__cssModules",Q]]);function ht(){const e=F(),t=h(),o=h();return q(),{email:t,password:o,handleLogin:async()=>(console.log(t.value),console.log(o.value),await e.push("/profile/"))}}const ft="_wrapper_gqhtr_20";var gt={wrapper:ft};const bt=m({setup(){const{email:e,password:t,handleLogin:o}=ht();return{email:e,password:t,handleLogin:o}}});function vt(e,t,o,c,r,a){const l=X;return i(),u("div",{class:n(e.$style.wrapper)},[d(l,{email:e.email,"onUpdate:email":t[0]||(t[0]=p=>e.email=p),password:e.password,"onUpdate:password":t[1]||(t[1]=p=>e.password=p),buttonText:"\u30ED\u30B0\u30A4\u30F3",onClick:e.handleLogin},null,8,["email","password","onClick"])],2)}const Y={};Y.$style=gt;var wt=$(bt,[["render",vt],["__cssModules",Y]]);function kt(){const e=F(),t=h(),o=h();return q(),{email:t,password:o,handleSignup:async()=>(console.log(t.value),console.log(o.value),await e.push("/profile/"))}}const Mt="_wrapper_gqhtr_20";var Ct={wrapper:Mt};const Dt=m({setup(){const{email:e,password:t,handleSignup:o}=kt();return{email:e,password:t,handleSignup:o}}});function Et(e,t,o,c,r,a){const l=X;return i(),u("div",{class:n(e.$style.wrapper)},[d(l,{email:e.email,"onUpdate:email":t[0]||(t[0]=p=>e.email=p),password:e.password,"onUpdate:password":t[1]||(t[1]=p=>e.password=p),buttonText:"\u30B5\u30A4\u30F3\u30A2\u30C3\u30D7",onClick:e.handleSignup},null,8,["email","password","onClick"])],2)}const J={};J.$style=Ct;var At=$(Dt,[["render",Et],["__cssModules",J]]);const Bt="_wrapper_1u1ik_20";var It={wrapper:Bt};const St=m({setup(){return{}}}),Ft=s("span",null,"\u540D\u524D",-1),qt=s("input",{type:"text"},null,-1),Lt=[Ft,qt],Nt=s("div",null,[s("span",null,"\u8077\u7A2E"),s("select",{name:"",id:""})],-1),Ot=s("div",null,[s("span",null,"\u81EA\u5DF1\u7D39\u4ECB"),s("textarea",{name:"",id:"",cols:"30",rows:"10"})],-1),Pt=s("span",null,"\u73FE\u5728\u52E4\u3081\u3066\u3044\u308B\u4F1A\u793E",-1),Rt=s("input",{type:"text"},null,-1),Vt=[Pt,Rt],zt=s("div",null,[s("span",null,"\u5B9F\u52D9\u7D4C\u9A13\u5E74\u6570"),s("select",{name:"",id:""})],-1),Gt=s("span",null,"facebook",-1),Ht=s("input",{type:"text"},null,-1),Tt=[Gt,Ht],Ut=s("span",null,"twitter",-1),jt=s("input",{type:"text"},null,-1),Wt=[Ut,jt],Kt=s("span",null,"github",-1),Qt=s("input",{type:"text"},null,-1),Xt=[Kt,Qt],Yt=s("span",null,"\u4F1A\u793E\u540D",-1),Jt=s("input",{type:"text"},null,-1),Zt=[Yt,Jt],xt=s("span",null,"\u5165\u793E\u65E5",-1),en=s("input",{type:"text"},null,-1),sn=[xt,en],tn=s("span",null,"\u9000\u793E\u65E5",-1),nn=s("input",{type:"text"},null,-1),on=[tn,nn],rn=s("div",null,[s("span",null,"\u4ED5\u4E8B\u5185\u5BB9"),s("textarea",{name:"",id:"",cols:"30",rows:"10"})],-1),ln=s("div",null,[s("span",null,"\u4F7F\u7528\u8A00\u8A9E"),s("select",{name:"",id:""})],-1),an=s("div",null,[s("span",null,"\u30DD\u30B8\u30B7\u30E7\u30F3"),s("select",{name:"",id:""})],-1);function un(e,t,o,c,r,a){return i(),u("div",{class:n(e.$style.wrapper)},[s("div",{class:n(e.$style.labelInput)},Lt,2),Nt,Ot,s("div",{class:n(e.$style.labelInput)},Vt,2),zt,s("div",null,[s("div",{class:n(e.$style.labelInput)},Tt,2),s("div",{class:n(e.$style.labelInput)},Wt,2),s("div",{class:n(e.$style.labelInput)},Xt,2)]),s("div",null,[s("div",{class:n(e.$style.labelInput)},Zt,2),s("div",{class:n(e.$style.labelInput)},sn,2),s("div",{class:n(e.$style.labelInput)},on,2),rn,ln,an])],2)}const Z={};Z.$style=It;var cn=$(St,[["render",un],["__cssModules",Z]]);const pn=[{path:"/",name:"Home",component:Te},{path:"/member",name:"Member",component:as},{path:"/member/:id",component:xs},{path:"/login",component:wt},{path:"/signup",component:At},{path:"/profile",component:cn}],dn=ce({history:pe(),routes:pn}),_n="_wrapper_2exak_20";var mn={wrapper:_n};const $n=m({setup(){return{}}}),yn=s("div",null,"R - World",-1),hn=[yn];function fn(e,t,o,c,r,a){return i(),u("footer",{class:n(e.$style.wrapper)},hn,2)}const x={};x.$style=mn;var ee=$($n,[["render",fn],["__cssModules",x]]);const gn="_app_16okz_1",bn="_mainContents_16okz_25";var vn={app:gn,mainContents:bn};const wn=m({name:"App",components:{OGlobalHeader:E,OGlobalFooter:ee}});function kn(e,t,o,c,r,a){const l=E,p=D("router-view"),y=ee;return i(),u(b,null,[d(l),d(p,{class:n(e.$style.mainContents)},null,8,["class"]),d(y)],64)}const se={};se.$style=vn;var Mn=$(wn,[["render",kn],["__cssModules",se]]);const Cn={apiKey:"AIzaSyB-qukIHon_UxbN3iBYggLpt3fmck_lHkw",authDomain:"toipptakosan11-71185.firebaseapp.com",databaseURL:"https://toipptakosan11-71185.firebaseio.com",projectId:"toipptakosan11-71185",storageBucket:"toipptakosan11-71185.appspot.com",messagingSenderId:"623468101161",appId:"1:623468101161:web:63273b946929fdb87aa638",measurementId:"G-C4131DCWTX"};de(Cn);_e(Mn).use(dn).mount("#app");
