+++ public/index.html
@@ -5,7 +5,7 @@
     <link rel="icon" href="/favicon.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>toipptakosan11</title>
-    <script type="module" crossorigin src="./assets/index.898c63c1.js"></script>
+    <script type="module" crossorigin src="./assets/index.91529c3a.js"></script>
     <link rel="modulepreload" href="./assets/vendor.b4423c4d.js">
     <link rel="stylesheet" href="./assets/index.6262914f.css">
   </head>
